CSS Regions build bot should archive and upload output files https://bugs.webkit.org/show_bug.cgi?id=65200
Until now the buildbots would only upload the result archive if the 'trigger' parameters was specified. The
upload was needed, because the testbots needed to download and use the archive. For CSS Regions we don't have
a testbot yet, but we want to save the archive on the server for manual testing. Added new parameter
in config.json, called "upload" that when set to "true" will force the upload to the buildmaster server.

The "features" argument is now taken into account when determining the name of the archive on the server. That
was needed in order to avoid the CSS Regions Mac buildbot overwrite the SnowLeopard Mac build.

Patch by Alexandru Chiculita <achicu@adobe.com> on 2011-08-02
Reviewed by Adam Roben.

* BuildSlaveSupport/build.webkit.org-config/config.json:
* BuildSlaveSupport/build.webkit.org-config/master.cfg:

Identifier: 81334@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92191 268f45cc-cd09-0410-ab3c-d52691b4dbfc