TiledDrawingArea: Handle update requests in the order they were received. https://bugs.webkit.org/show_bug.cgi?id=64365
Reviewed by Andreas Kling.

TiledDrawingAreaProxy::createTiles() uses the distance of tiles to the
center of the viewport to decide which tile should be rendered first.
This logic is useless if the requests are not handled in the same order
as they were received.
Now use a list instead of a map to hold pending tile update requests.

* WebProcess/WebPage/TiledDrawingArea.cpp:
(WebKit::TiledDrawingArea::tileUpdateTimerFired):
(WebKit::TiledDrawingArea::cancelTileUpdate):
(WebKit::TiledDrawingArea::requestTileUpdate):
* WebProcess/WebPage/TiledDrawingArea.h:

Identifier: 81341@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92198 268f45cc-cd09-0410-ab3c-d52691b4dbfc