Chromium Mac: Make sure scrollbars flash when web page loads https://bugs.webkit.org/show_bug.cgi?id=65586
On slow web pages scrollbars wouldn't flash when the page was done loading. The problem was that we were flashing the scrollbar 0.1 second after the load operation had started. If the page was slow to load then we might not have scrollbars to flash. To work around this I added an extra check to make sure that we had indeed finished loading the page before flashing the scrollbars.

Patch by Sailesh Agrawal <sail@chromium.org> on 2011-08-03
Reviewed by James Robinson.

* platform/chromium/ScrollAnimatorChromiumMac.mm:
(WebCore::ScrollAnimatorChromiumMac::initialScrollbarPaintTimerFired):

Identifier: 81444@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92316 268f45cc-cd09-0410-ab3c-d52691b4dbfc