Attempt to fix the build after r92538.
Like all of the recent "track"-related work the new files in r92538 appear to have
been added to completely random places in the Xcode project. However, this time
the file paths were marked as being relative to their containing group resulting in
Xcode being unable to find the files on disk.

I've attempted to clean up all of the "track"-related mess in the Xcode project.
The location of the files in the project now reflect their locations on disk, and
are all correctly marked as being relative to their containing group as is our
convention.

* WebCore.xcodeproj/project.pbxproj:

Identifier: 81637@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92542 268f45cc-cd09-0410-ab3c-d52691b4dbfc