Test how SpeechInputResultList behaves with not enough arguments https://bugs.webkit.org/show_bug.cgi?id=65812
Reviewed by Sam Weinig.

SpeechInputResultList is only exposed via the x-webkit-speech
attribute, which means there is little compatibility risk in aligning
our too-few-arguments behavior with the specification.  There's a
question about whether all these "fake array" objects should work
consistently, but we'll probably want the "fake array" objects we
create in the future to match the spec, so even if we're consistent
now, we'll end up not being consistent in the future.

* fast/speech/speech-input-result-list-not-enough-arguments-expected.txt: Added.
* fast/speech/speech-input-result-list-not-enough-arguments.html: Added.


Identifier: 81900@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92828 268f45cc-cd09-0410-ab3c-d52691b4dbfc