[Qt] autotests shouldn't require config.h in the style check https://bugs.webkit.org/show_bug.cgi?id=65986
Remove the include rules from the style-check for Qt tests.
Since for the most part it's just a matter of using the
API provided by QtWebKit, enforcing the WebCore set of rules
only seems to make sense in the case of MIMESniffing.

Patch by Pierre Rossi <pierre.rossi@gmail.com> on 2011-08-11
Reviewed by Benjamin Poulain.

* Scripts/webkitpy/style/checker.py:
* Scripts/webkitpy/style/checker_unittest.py:

Identifier: 81918@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92847 268f45cc-cd09-0410-ab3c-d52691b4dbfc