The PageSerializer should always use absolute URL https://bugs.webkit.org/show_bug.cgi?id=66006
Patch by Benjamin Poulain <benjamin@webkit.org> on 2011-08-11
Reviewed by Adam Barth.

Page serializer resolves every URL for storing the resources.
The MarkupAccumulator used there should use the same rules in order to
create a valid file.

No new tests because the test infrastructure for this
does not exist with layout tests and the Chromium unit test for
the serializer are disabled.

* page/PageSerializer.cpp:
(WebCore::SerializerMarkupAccumulator::SerializerMarkupAccumulator):

Identifier: 81948@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92883 268f45cc-cd09-0410-ab3c-d52691b4dbfc