IndexedDB: Overwriting key in unique index should be possible https://bugs.webkit.org/show_bug.cgi?id=65993
Reviewed by Tony Chang.

Source/WebCore:

It should be possible to overwrite an object store record even if
there is a derived key for that record in an index with the unique flag set.

* storage/IDBBackingStore.h:
* storage/IDBIndexBackendImpl.cpp:
(WebCore::IDBIndexBackendImpl::addingKeyAllowed):
* storage/IDBIndexBackendImpl.h:
* storage/IDBLevelDBBackingStore.cpp:
(WebCore::IDBLevelDBBackingStore::keyExistsInIndex):
* storage/IDBLevelDBBackingStore.h:
* storage/IDBObjectStoreBackendImpl.cpp:
(WebCore::IDBObjectStoreBackendImpl::putInternal):
* storage/IDBSQLiteBackingStore.cpp:
(WebCore::IDBSQLiteBackingStore::keyExistsInIndex):
* storage/IDBSQLiteBackingStore.h:

LayoutTests:

Test that it's possible to overwrite an object store record even if there
is a derived key for that record in an index with the unique flag set.

* storage/indexeddb/index-unique-expected.txt:
* storage/indexeddb/index-unique.html:

Identifier: 81979@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92950 268f45cc-cd09-0410-ab3c-d52691b4dbfc